feat(FIJI-2344): [Display Events] move common implementation from 2926, it is reviewed(https://git.ringcentral.com/Fiji/Fiji/merge_requests/1052/diffs)

+++ packages/sdk/src/module/item/controller/ItemServiceController.ts
@@ -11,7 +11,7 @@ import { ControllerBuilder } from '../../../framework/controller/impl/Controller
 import { Item } from '../entity';
 import { Api } from '../../../api';
 import { daoManager, ItemDao } from '../../../dao';
-import { GlipTypeUtil, TypeDictionary } from '../../../utils';
+import { GlipTypeUtil } from '../../../utils';
 import { IItemService } from '../service/IItemService';
 import { ItemQueryOptions, ItemFilterFunction } from '../types';
 
@@ -117,18 +117,11 @@ class ItemServiceController {
   async handleSanitizedItems(incomingItems: Item[]) {
     const typeItemsMap: Map<number, Item[]> = new Map();
     incomingItems.forEach((item: Item) => {
-      const type = GlipTypeUtil.extractTypeId(item.id);
-      switch (type) {
-        case TypeDictionary.TYPE_ID_FILE:
-          typeItemsMap.has(TypeDictionary.TYPE_ID_FILE)
-            ? (typeItemsMap.get(TypeDictionary.TYPE_ID_FILE) as Item[]).push(
-                item,
-              )
-            : typeItemsMap.set(TypeDictionary.TYPE_ID_FILE, [item]);
-          break;
-        default:
-          break;
-      }
+      this._saveItemsToTypeIdMap(
+        GlipTypeUtil.extractTypeId(item.id),
+        typeItemsMap,
+        item,
+      );
     });
 
     typeItemsMap.forEach(
@@ -138,8 +131,21 @@ class ItemServiceController {
     );
   }
 
+  private _saveItemsToTypeIdMap(
+    typeId: number,
+    typeItemsMap: Map<number, Item[]>,
+    item: Item,
+  ) {
+    typeItemsMap.has(typeId)
+      ? (typeItemsMap.get(typeId) as Item[]).push(item)
+      : typeItemsMap.set(typeId, [item]);
+  }
+
   private async _updateSanitizedItems(typeId: number, items: Item[]) {
     const subItemService = this.getSubItemService(typeId);
+    if (!subItemService) {
+      return;
+    }
     const deactivatedItems = items.filter(item => item.deactivated);
     deactivatedItems.forEach((item: Item) => {
       subItemService.deleteItem(item.id);
